Make the coverage target a whole number of miles
`backbone.coverage_target_miles` was a float because that is how it was
written the day the knob was added, not because the model has a fraction
of a mile to express. Its one use is the growth stop test
`worst <= target_miles`, where `worst` is a great-circle distance
standing in for a last-mile build nobody has measured: the error between
the number and the world is tens of miles, so a target of `400.5` states
a resolution the design does not have.

Read the key with `_required_int` and type the `Tuning` field `int`,
which leaves `_required_float` with no caller, so remove it. The parser
change and the configs land together because YAML reads `400.0` as a
float and `_required_int` rejects one -- splitting them leaves `main`
unparseable in between.

Closes #19

+++ src/api/endpoints/tenants/wan/post/lambdas/synthesizer/model.py
@@ -84,7 +84,9 @@ class Tuning:
 
     compass_sector_count: int = 8  # compass sectors used to score a backbone node's link spread
     backbone_mesh_degree: int = 3  # other backbone nodes each one wires to (backbone-mesh-degree)
-    backbone_coverage_target_miles: float = 600.0  # grow backbone until every demand is this near
+    # grow backbone until every demand is this near; whole miles, since the great-circle
+    # haul it is compared against stands in for an unmeasured last-mile build
+    backbone_coverage_target_miles: int = 600
     access_backbone_links: int = 2  # backbone nodes each demand vertex homes to
     # how much memory the backbone enumeration may spend
     search_memory_budget: SearchMemoryBudget = field(default_factory=SearchMemoryBudget)
